Add URL Prefix argument to site builder
The build is now working but on GitHub pages the site is rooted at the
path `/attack-sync/` not `/`, which is creating broken links for just
about every URL we construct. I added some commmand line arguments
to pass in a URL prefix so that we can construct the correct absolute
URLs.