Fix version number lookup.
The logic was using the release_info module from mitreattack-python, but
that does lookups based on file hash and some of the releases have
identical hashes (i.e. no changes), so this would cause, for example, a
page for mobile v11.0 to call itself v10.1 instead. This update uses the
version numbers that the user passes in.