Move the JSON download link
It was displayed on the changelog landing page, but that was a
confusing spot because:

- you have to know how to get to the landing page before you can find
  the download link
- the landing page is scoped to a single domain, whereas the json
  download includes all domains

The updated solution is to add download menu to the navbar. Since the
bar is already crowd, the download menu uses an icon instead of text on
lg viewports. On small viewports (e.g. phone/tablet) that have the
hamburger menu, the icon switches to the text "Downloads" so that its
more intelligible.